Contains classes for Toolbars.
TBContainer sets up a container for toolbar widgets which can be right
clicked to list available toolbars.

Toolbar sets up a toolbar widget which holds a handle, optional label, and
a series of buttons. The handle can be right clicked to open a menu of
available options for the toolbar including moving it within the container,
tearing it off the container and into a toplevel window, removing it from
view and toggling whether the label is shown.

TBButton defines what a button for the toolbars looks like.

Known bugs:
1. If a toolbar is too large to be completely displayed, it is created and
then destroyed leading to flickering.
2. If a toolbar which is too large is tried to be displayed twice, all other
toolbars will disappear and the toolbar which is too large will take their
place.